+++ backend/src/main/java/com/techeer/carpool/domain/comment/dto/CommentRequest.java
@@ -7,7 +7,7 @@ import lombok.NoArgsConstructor;
 
 @Getter
 @NoArgsConstructor
-public class CommentCreateRequest {
+public class CommentRequest {
 
     @NotBlank(message = "댓글 내용을 입력해주세요.")
     @Size(max = 500, message = "댓글은 500자 이내로 입력해주세요.")
