--- backend/src/main/java/com/techeer/carpool/domain/comment/service/CommentReadService.java
@@ -1,46 +1,0 @@
-package com.techeer.carpool.domain.comment.service;
-
-import com.techeer.carpool.domain.comment.dto.CommentResponse;
-import com.techeer.carpool.domain.comment.entity.Comment;
-import com.techeer.carpool.domain.comment.repository.CommentRepository;
-import com.techeer.carpool.domain.member.entity.Member;
-import com.techeer.carpool.domain.member.repository.MemberRepository;
-import com.techeer.carpool.domain.post.repository.PostRepository;
-import com.techeer.carpool.global.exception.CarpoolException;
-import com.techeer.carpool.global.exception.ErrorCode;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-import java.util.List;
-import java.util.Map;
-import java.util.Set;
-import java.util.stream.Collectors;
-
-@Service
-@RequiredArgsConstructor
-public class CommentReadService {
-
-    private final CommentRepository commentRepository;
-    private final PostRepository postRepository;
-    private final MemberRepository memberRepository;
-
-    @Transactional(readOnly = true)
-    public List<CommentResponse> getCommentsByPostId(Long postId) {
-        postRepository.findByIdAndDeletedFalse(postId)
-                .orElseThrow(() -> new CarpoolException(ErrorCode.POST_NOT_FOUND));
-
-        List<Comment> comments = commentRepository.findByPostIdAndDeletedFalseOrderByCreatedAtAsc(postId);
-
-        Set<Long> memberIds = comments.stream()
-                .map(Comment::getMemberId)
-                .collect(Collectors.toSet());
-
-        Map<Long, String> nicknameMap = memberRepository.findAllById(memberIds).stream()
-                .collect(Collectors.toMap(Member::getId, Member::getNickname));
-
-        return comments.stream()
-                .map(c -> CommentResponse.from(c, nicknameMap.getOrDefault(c.getMemberId(), "알 수 없음")))
-                .collect(Collectors.toList());
-    }
-}
